requirements.txt
pyyaml
mkdocs-material